refactor: move metadata storage class to root
