+++ src/storage.ts
@@ -1,4 +1,4 @@
-import { MetadataStorage } from './metadata/MetadataStorage';
+import { MetadataStorage } from './MetadataStorage';
 
 /**
  * Default metadata storage is used as singleton and can be used to storage all metadatas.
